Postopedia user roles - Working but... (#4)
* Postopedia User Roles, first pass default user working

* Postopedia User Roles, second pass, getting ActiveRecord::PendingMigrationError

* Postopedia User Roles, third pass, close to fully functional

* Postopedia User Roles, seeded

* Postopedia User Roles working

+++ app/controllers/wikis_controller.rb
@@ -1,19 +1,24 @@
 class WikisController < ApplicationController
+  after_action :verify_authorized, except: [:index]
+
   def index
     @wikis = Wiki.all
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @wiki = Wiki.new(wiki_params)
     @wiki.user = current_user
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was saved."
@@ -27,6 +32,7 @@ class WikisController < ApplicationController
   def update
     @wiki = Wiki.find(params[:id])
     @wiki.assign_attributes(wiki_params)
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was updated."
@@ -39,14 +45,16 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-      redirect_to [@wiki]
+      redirect_to wikis_path
     else
       flash[:error] = "There was an error deleting the wiki."
       render :show
